Change implicit casts to explicit.  No functional changes.

+++ libjsmn/jsmn.c
@@ -41,7 +41,7 @@ static jsmnerr_t jsmn_parse_primitive(jsmn_parser *parser, const char *js,
 	jsmntok_t *token;
 	int start;
 
-	start = parser->pos;
+	start = (int)parser->pos;
 
 	for (; parser->pos < len && js[parser->pos] != '\0'; parser->pos++) {
 		switch (js[parser->pos]) {
@@ -54,7 +54,7 @@ static jsmnerr_t jsmn_parse_primitive(jsmn_parser *parser, const char *js,
 				goto found;
 		}
 		if (js[parser->pos] < 32 || js[parser->pos] >= 127) {
-			parser->pos = start;
+			parser->pos = (unsigned int)start;
 			return JSMN_ERROR_INVAL;
 		}
 	}
@@ -71,10 +71,10 @@ found:
 	}
 	token = jsmn_alloc_token(parser, tokens, num_tokens);
 	if (token == NULL) {
-		parser->pos = start;
+		parser->pos = (unsigned int)start;
 		return JSMN_ERROR_NOMEM;
 	}
-	jsmn_fill_token(token, JSMN_PRIMITIVE, start, parser->pos);
+	jsmn_fill_token(token, JSMN_PRIMITIVE, start, (int)parser->pos);
 #ifdef JSMN_PARENT_LINKS
 	token->parent = parser->toksuper;
 #endif
@@ -89,7 +89,7 @@ static jsmnerr_t jsmn_parse_string(jsmn_parser *parser, const char *js,
 		size_t len, jsmntok_t *tokens, size_t num_tokens) {
 	jsmntok_t *token;
 
-	int start = parser->pos;
+	int start = (int)parser->pos;
 
 	parser->pos++;
 
@@ -104,10 +104,11 @@ static jsmnerr_t jsmn_parse_string(jsmn_parser *parser, const char *js,
 			}
 			token = jsmn_alloc_token(parser, tokens, num_tokens);
 			if (token == NULL) {
-				parser->pos = start;
+				parser->pos = (unsigned int)start;
 				return JSMN_ERROR_NOMEM;
 			}
-			jsmn_fill_token(token, JSMN_STRING, start+1, parser->pos);
+			jsmn_fill_token(token, JSMN_STRING, start+1,
+                                        (int)parser->pos);
 #ifdef JSMN_PARENT_LINKS
 			token->parent = parser->toksuper;
 #endif
@@ -127,25 +128,25 @@ static jsmnerr_t jsmn_parse_string(jsmn_parser *parser, const char *js,
 				case 'u':
 					parser->pos++;
 					for(i = 0; i < 4 && parser->pos < len && js[parser->pos] != '\0'; i++) {
-						/* If it isn't a hex character we have an error */
-						if(!((js[parser->pos] >= 48 && js[parser->pos] <= 57) || /* 0-9 */
-									(js[parser->pos] >= 65 && js[parser->pos] <= 70) || /* A-F */
-									(js[parser->pos] >= 97 && js[parser->pos] <= 102))) { /* a-f */
-							parser->pos = start;
-							return JSMN_ERROR_INVAL;
-						}
-						parser->pos++;
+					    /* If it isn't a hex character we have an error */
+					    if(!((js[parser->pos] >= 48 && js[parser->pos] <= 57) || /* 0-9 */
+								    (js[parser->pos] >= 65 && js[parser->pos] <= 70) || /* A-F */
+								    (js[parser->pos] >= 97 && js[parser->pos] <= 102))) { /* a-f */
+						parser->pos = (unsigned int)start;
+						return JSMN_ERROR_INVAL;
+					    }
+					    parser->pos++;
 					}
 					parser->pos--;
 					break;
 				/* Unexpected symbol */
 				default:
-					parser->pos = start;
+					parser->pos = (unsigned int)start;
 					return JSMN_ERROR_INVAL;
 			}
 		}
 	}
-	parser->pos = start;
+	parser->pos = (unsigned int)start;
 	return JSMN_ERROR_PART;
 }
 
@@ -180,8 +181,8 @@ jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
 #endif
 				}
 				token->type = (c == '{' ? JSMN_OBJECT : JSMN_ARRAY);
-				token->start = parser->pos;
-				parser->toksuper = parser->toknext - 1;
+				token->start = (int)parser->pos;
+				parser->toksuper = (int)parser->toknext - 1;
 				break;
 			case '}': case ']':
 				if (tokens == NULL)
@@ -194,12 +195,12 @@ jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
 				token = &tokens[parser->toknext - 1];
 				for (;;) {
 					if (token->start != -1 && token->end == -1) {
-						if (token->type != type) {
-							return JSMN_ERROR_INVAL;
-						}
-						token->end = parser->pos + 1;
-						parser->toksuper = token->parent;
-						break;
+					    if (token->type != type) {
+						    return JSMN_ERROR_INVAL;
+					    }
+					    token->end = (int)parser->pos + 1;
+					    parser->toksuper = token->parent;
+					    break;
 					}
 					if (token->parent == -1) {
 						break;
@@ -239,7 +240,7 @@ jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
 			case '\t' : case '\r' : case '\n' : case ' ':
 				break;
 			case ':':
-				parser->toksuper = parser->toknext - 1;
+				parser->toksuper = (int)parser->toknext - 1;
 				break;
 			case ',':
 				if (tokens != NULL &&
@@ -291,7 +292,7 @@ jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
 		}
 	}
 	if (tokens != NULL) {
-		for (i = parser->toknext - 1; i >= 0; i--) {
+		for (i = (int)parser->toknext - 1; i >= 0; i--) {
 			/* Unmatched opened object or array */
 			if (tokens[i].start != -1 && tokens[i].end == -1) {
 				return JSMN_ERROR_PART;
